extract world interactions
