+++ crates/client/src/app/input/mod.rs
@@ -8,5 +8,6 @@ pub mod look_mode;
 pub mod multi_action;
 pub mod npc_interaction;
 pub mod splash;
+pub mod world_action;
 pub mod world_build;
 pub mod world_movement;
